增强子命令表达方式
